feat: Add support for multiple views
This adds the support to have multiple views when printing tasks to the
CLI. This is also supported by giving configuration options. There is
always one table, but the table can be separated automatically based on
the tasks' filter or on user-defined named filters.

Colours can be applied to each section.
Each section filter will further split the tasks matched by the report
filter. All tasks matched by the report filter (+ user query optionally)
will be shown in the table.